requirements.txt
Django~=3.0.6
djangorestframework~=3.11.0
psycopg2-binary~=2.8.5
gunicorn~=20.0.4
django-cors-headers~=3.3.0
Pillow~=7.1.2
requests~=2.23.0
pytest~=6.1.2
opencv-python~=4.4.0.44
numpy~=1.18.5
